+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.10] - 2024-09-04
+
+### Changed
+
+- The `--timestamp` commandline option was replaced with `--title-caption`. This allows the text below the document title (on the ToC page) to be anything, not just a string with " GMT" attached.
+- The default value of both `--title-caption` and `--output` is now printed in the help output.
+- The application now exits with an error code (and not just logs errors) if an error occurred.
+
 ## [1.0.9] - 2024-09-04
 
 ### Added